refactor: Move `XmlRelationship` conversion to a separate sub-module and dedicated traits.
