+++ src/serde/xml/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod xml_relationship;
 pub(crate) mod xml_variable;
 
+pub(crate) use xml_relationship::XmlRelationship;
 pub(crate) use xml_variable::XmlVariable;
